refactor(action): move async action name mapping into constants folder
