+++ src/actions/venues.ts
@@ -1,14 +1,10 @@
-import { ActionType, createAsyncAction, getType } from 'typesafe-actions'
+import { ActionType, createAsyncAction } from 'typesafe-actions'
 import { NLikes } from '../../types/likes.d'
 import { NMiniVenue } from '../../types/miniVenue.d'
 import { NRecommendedPlaces } from '../../types/recommendedPlaces'
 import { NRequest } from '../../types/request.d'
 import { NVenue } from '../../types/venue.d'
 import { NVenuesCategories } from '../../types/venuesCategories'
-
-interface IAsyncActionNameMapping {
-  [actionType: string]: string
-}
 
 export const getVenuesCategoriesAsync = createAsyncAction(
   'GET_VENUES_CATEGORIES_REQUEST',
@@ -61,34 +57,6 @@ export const getVenuesSuggestCompletionAsync = createAsyncAction(
   Error,
   string
 >()
-
-export const ASYNC_ACTION_NAME_MAPPING: IAsyncActionNameMapping = {
-  // getVenuesCategoriesAsync
-  [getType(getVenuesCategoriesAsync.request)]: 'getVenuesCategoriesAsync',
-  [getType(getVenuesCategoriesAsync.success)]: 'getVenuesCategoriesAsync',
-  [getType(getVenuesCategoriesAsync.failure)]: 'getVenuesCategoriesAsync',
-  [getType(getVenuesCategoriesAsync.cancel)]: 'getVenuesCategoriesAsync',
-  // getVenuesExploreAsync
-  [getType(getVenuesExploreAsync.request)]: 'getVenuesExploreAsync',
-  [getType(getVenuesExploreAsync.success)]: 'getVenuesExploreAsync',
-  [getType(getVenuesExploreAsync.failure)]: 'getVenuesExploreAsync',
-  [getType(getVenuesExploreAsync.cancel)]: 'getVenuesExploreAsync',
-  // getVenuesLikesAsync
-  [getType(getVenuesLikesAsync.request)]: 'getVenuesLikesAsync',
-  [getType(getVenuesLikesAsync.success)]: 'getVenuesLikesAsync',
-  [getType(getVenuesLikesAsync.failure)]: 'getVenuesLikesAsync',
-  [getType(getVenuesLikesAsync.cancel)]: 'getVenuesLikesAsync',
-  // getVenuesSearchAsync
-  [getType(getVenuesSearchAsync.request)]: 'getVenuesSearchAsync',
-  [getType(getVenuesSearchAsync.success)]: 'getVenuesSearchAsync',
-  [getType(getVenuesSearchAsync.failure)]: 'getVenuesSearchAsync',
-  [getType(getVenuesSearchAsync.cancel)]: 'getVenuesSearchAsync',
-  // getVenuesTrendingAsync
-  [getType(getVenuesTrendingAsync.request)]: 'getVenuesTrendingAsync',
-  [getType(getVenuesTrendingAsync.success)]: 'getVenuesTrendingAsync',
-  [getType(getVenuesTrendingAsync.failure)]: 'getVenuesTrendingAsync',
-  [getType(getVenuesTrendingAsync.cancel)]: 'getVenuesTrendingAsync',
-}
 
 export const venuesActions = {
   getVenuesCategoriesAsync,
